refactor to move kubectl.cmd.Factory to kubect/cmd/util
